+++ piq/pieapp.py
@@ -161,8 +161,8 @@ class PieAPP(_Loss):
                           'The input images were converted to RGB by copying the grey channel 3 times.')
 
         self.model.to(device=x.device)
-        x_features, x_weights = self.get_features(x)
-        y_features, y_weights = self.get_features(y)
+        x_features, x_weights = self.get_features(x.float())
+        y_features, y_weights = self.get_features(y.float())
 
         distances, weights = self.model.compute_difference(
             y_features - x_features,
